complied with unit test naming coventions, changed greeting to group name
